fix automation actions for A_isf

+++ plugins/automation/src/main/kotlin/app/aaps/plugins/automation/elements/InputIobTH.kt
@@ -8,17 +8,17 @@ import java.text.DecimalFormat
 class InputIobTH() : Element {
 
     // @Inject lateinit var sp: SP
-    var value = 100.0 // not working on app start: sp.getDouble(R.string.key_bgAccel_ISF_weight, 0.0)
+    var value = 100 // not working on app start: sp.getDouble(R.string.key_iob...., 100.0)
 
-    constructor(another: Double) : this() {
+    constructor(another: Int) : this() {
         value = another
     }
 
     override fun addToLayout(root: LinearLayout) {
         root.addView(
             NumberPicker(root.context, null).also {
-                it.setParams(value, 10.0, 100.0, 5.0, DecimalFormat("0"), false, root.findViewById(app.aaps.core.ui.R.id.ok))
-                it.setOnValueChangedListener { v: Double -> value = v }
+                it.setParams(value.toDouble(), 10.0, 100.0, 5.0, DecimalFormat("0"), false, root.findViewById(app.aaps.core.ui.R.id.ok))
+                it.setOnValueChangedListener { v: Double -> value = v.toInt() }
                 it.gravity = Gravity.CENTER_HORIZONTAL
             })
     }
